# docs/private_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# Introduction (OpenAPI Specification)\nCinnabar Private REST API merupakan API private yang tersedia untuk umum dan dapat digunakan secara secara berbayar. Dokumentasi API ini merupakan dokumentasi Cinnabar REST API untuk privat. Apabila anda ingin melihat dokumentasi Cinnabar REST API yang privat, silakan menuju tautan [Cinnabar Public REST API -  Documentation](https://cinnabar.icaksh.my.id/public/docs).\nAPI ini didokumentasikan dalam bentuk **OpenAPI format** dan didasarkan pada [swagger.io](https://swagger.io). Halaman dokumentasi ini digenerate dengan menggunakan [swaggo](https://github.com/swaggo/swag) dan dipublikasikan menggunakan [redoc](https://github.com/Redocly/redoc).\n# Cross-Origin Resource Sharing\nAPI ini memiliki fitur Cross-Origin Resource Sharing (CORS) yang diterapkan sesuai dengan [spesifikasi W3C](https://www.w3.org/TR/cors/) dan memungkinkan komunikasi lintas domain. Semua response bersifat publik dan dapat diakses oleh semua orang dari aplikasi apapun ataupun website apapun.\n# Information\nAPI ini masih dalam tahap pengembangan. Apabila anda ingin memberikan saran terhadap pengembangan API ini, silakan hubungi alamat pos-el yang telah tersedia. Jangan sungkan untuk menghubungi alamat pos-el jika terjadi error",
        "title": "Cinnabar Private REST API",
        "contact": {
            "name": "Cinnabar REST API Support",
            "url": "https://cinnabar.icaksh.my.id/",
            "email": "me@icaksh.my.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "BETA",
        "x-logo": {
            "altText": "Cinnabar REST API",
            "backgroundColor": "#ffffff",
            "url": "https://cinnabar.icaksh.my.id/view/logo.png"
        }
    },
    "host": "cinnabar.icaksh.my.id",
    "basePath": "/public",
    "paths": {
        "/api/info": {
            "post": {
                "security": [
                    {
                        "apikey": []
                    }
                ],
                "description": "Check your API Information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basic"
                ],
                "summary": "REST - Api Information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TApiKey"
                        }
                    }
                }
            }
        },
        "/api/test": {
            "post": {
                "security": [
                    {
                        "apikey": []
                    }
                ],
                "description": "Cinnabar REST API JSON Format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basic"
                ],
                "summary": "REST - Response",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TSuccessResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.TApiKey": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "quota": {
                    "type": "integer"
                }
            }
        },
        "model.TSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "apikey": {
            "description": "For this sample, you can use the api key `special-key` to test the authorization filters.",
            "type": "apiKey",
            "name": "apikey",
            "in": "header"
        }
    },
    "tags": [
        {
            "name": "basic"
        }
    ],
    "x-tagGroups": [
        {
            "name": "API",
            "tags": [
                "basic"
            ]
        }
    ]
}